requirements.txt
streamlit
openai
faiss-cpu
numpy
requests
python-dotenv